Create MwilliamsQuiz
The code needs to be commented and the necessary questions and pictures added accordingly, for the images to show in bluej the pictures has to be within the same folder as the project and it classes it self, best if u download "PNG" pictures in hd. Replace the name of the file within the pictures parameter and run it as necessary to see it.